[Update] build tree function [New] add evaluate function (output correct result if simple calculation )

+++ src/cal/Operand.java
@@ -3,7 +3,7 @@ package cal;
 public class Operand extends TreeNode {
 	double val;
 	
-	public Operand(int value) {
+	public Operand(double value) {
 		val = value;
 		left = null;
 		right = null;
